+++ processor/src/test/fixtures/expected/com/example/prop/React4j_NonnullContextPropModel.java
@@ -1,0 +1,148 @@
+package com.example.prop;
+
+import arez.Disposable;
+import arez.annotations.ArezComponent;
+import arez.annotations.Feature;
+import java.util.Objects;
+import javax.annotation.Generated;
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import jsinterop.annotations.JsConstructor;
+import jsinterop.base.Js;
+import jsinterop.base.JsPropertyMap;
+import org.realityforge.braincheck.Guards;
+import react4j.React;
+import react4j.ReactNode;
+import react4j.internal.OnComponentWillUnmount;
+import react4j.internal.OnShouldComponentUpdate;
+import react4j.internal.ViewConstructorFunction;
+
+@SuppressWarnings("Arez:UnnecessaryAllowEmpty")
+@ArezComponent(
+    name = "NonnullContextPropModel",
+    disposeNotifier = Feature.DISABLE,
+    dagger = Feature.DISABLE,
+    sting = Feature.DISABLE,
+    allowEmpty = true
+)
+@Generated("react4j.processor.React4jProcessor")
+abstract class React4j_NonnullContextPropModel extends NonnullContextPropModel {
+  @Nonnull
+  private final react4j.internal.NativeView $$react4j$$_nativeView;
+
+  React4j_NonnullContextPropModel(
+      @Nonnull final react4j.internal.NativeView $$react4j$$_nativeView) {
+    this.$$react4j$$_nativeView = Objects.requireNonNull( $$react4j$$_nativeView );
+  }
+
+  @Nonnull
+  private static ViewConstructorFunction getConstructorFunction() {
+    final ViewConstructorFunction viewConstructor = ( React.shouldStoreDebugDataAsState() || React.shouldValidateInputValues() ) ? NativeView::new : LiteNativeView::new;
+    if ( React.enableViewNames() ) {
+      Js.asPropertyMap( viewConstructor ).set( "displayName", "NonnullContextPropModel" );
+    }
+    return viewConstructor;
+  }
+
+  @Nonnull
+  @Override
+  String getMyProp() {
+    return $$react4j$$_nativeView.inputs().getAsAny( Inputs.myProp ).asString();
+  }
+
+  private void $$react4j$$_validateInputValues(@Nonnull final JsPropertyMap<Object> inputs) {
+    final Object raw$myProp = inputs.get( Inputs.myProp );
+    if ( React.shouldCheckInvariants() ) {
+      Guards.apiInvariant( () -> null != raw$myProp, () -> "Context value of type java.lang.String is missing when constructing view named 'NonnullContextPropModel'. Ensure a parent view publishes the value to the context." ) ;
+    }
+  }
+
+  private boolean $$react4j$$_shouldComponentUpdate(
+      @Nullable final JsPropertyMap<Object> nextInputs) {
+    assert null != nextInputs;
+    if ( React.shouldValidateInputValues() ) {
+      $$react4j$$_validateInputValues( nextInputs );
+    }
+    final JsPropertyMap<Object> inputs = $$react4j$$_nativeView.inputs();
+    if ( !Js.isTripleEqual( inputs.get( Inputs.myProp ), nextInputs.get( Inputs.myProp ) ) ) {
+      return true;
+    }
+    return false;
+  }
+
+  private void $$react4j$$_componentWillUnmount() {
+    ((Arez_React4j_NonnullContextPropModel) this).dispose();
+  }
+
+  @Nullable
+  ReactNode $$react4j$$_render() {
+    assert Disposable.isNotDisposed( this );
+    return render();
+  }
+
+  static final class Factory {
+    @Nonnull
+    static final ViewConstructorFunction TYPE = getConstructorFunction();
+  }
+
+  static final class Inputs {
+    static final String myProp = React.shouldMinimizeInputKeys() ? "a" : "myProp";
+  }
+
+  private static final class LiteNativeView extends react4j.internal.NativeView implements OnShouldComponentUpdate {
+    @Nonnull
+    private final React4j_NonnullContextPropModel $$react4j$$_view;
+
+    @JsConstructor
+    LiteNativeView(@Nullable final JsPropertyMap<Object> inputs) {
+      super( inputs );
+      $$react4j$$_view = new Arez_React4j_NonnullContextPropModel( this );
+      if ( React.shouldValidateInputValues() ) {
+        assert null != inputs;
+        $$react4j$$_view.$$react4j$$_validateInputValues( inputs );
+      }
+    }
+
+    @Override
+    public final boolean shouldComponentUpdate(@Nonnull final JsPropertyMap<Object> nextInputs) {
+      return $$react4j$$_view.$$react4j$$_shouldComponentUpdate( nextInputs );
+    }
+
+    @Override
+    @Nullable
+    public final ReactNode render() {
+      return $$react4j$$_view.$$react4j$$_render();
+    }
+  }
+
+  private static final class NativeView extends react4j.internal.NativeView implements OnShouldComponentUpdate, OnComponentWillUnmount {
+    @Nonnull
+    private final React4j_NonnullContextPropModel $$react4j$$_view;
+
+    @JsConstructor
+    NativeView(@Nullable final JsPropertyMap<Object> inputs) {
+      super( inputs );
+      $$react4j$$_view = new Arez_React4j_NonnullContextPropModel( this );
+      if ( React.shouldValidateInputValues() ) {
+        assert null != inputs;
+        $$react4j$$_view.$$react4j$$_validateInputValues( inputs );
+      }
+    }
+
+    @Override
+    public final boolean shouldComponentUpdate(@Nonnull final JsPropertyMap<Object> nextInputs) {
+      return $$react4j$$_view.$$react4j$$_shouldComponentUpdate( nextInputs );
+    }
+
+    @Override
+    public final void componentWillUnmount() {
+      $$react4j$$_view.$$react4j$$_componentWillUnmount();
+    }
+
+    @Override
+    @Nullable
+    public final ReactNode render() {
+      return $$react4j$$_view.$$react4j$$_render();
+    }
+  }
+}