+++ processor/src/test/fixtures/input/com/example/prop/QualifiedNonnullContextPropModel.java
@@ -1,0 +1,23 @@
+package com.example.prop;
+
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import react4j.ReactNode;
+import react4j.annotations.Input;
+import react4j.annotations.Render;
+import react4j.annotations.View;
+
+@View
+abstract class QualifiedNonnullContextPropModel
+{
+  @Input( source = Input.Source.CONTEXT, qualifier = "my-proppo" )
+  @Nonnull
+  abstract String getMyProp();
+
+  @Nullable
+  @Render
+  ReactNode render()
+  {
+    return null;
+  }
+}